+++ src/os2datascanner/projects/report/reportapp/views/report_views.py
@@ -41,7 +41,9 @@ from os2datascanner.engine2.rules.links_follow import LinksFollowRule
 from os2datascanner.engine2.rules.rule import Sensitivity
 from os2datascanner.engine2.rules.wordlists import OrderedWordlistRule
 from os2datascanner.engine2.rules.dict_lookup import EmailHeaderRule
-from .view_utils import handle_report, delete_email
+
+from .utilities.document_report_utilities import handle_report
+from .utilities.msgraph_utilities import delete_email
 from ..models.documentreport import DocumentReport
 from ...organizations.models.account import Account
 from ...organizations.models.aliases import AliasType
@@ -237,6 +239,8 @@ class UserReportView(ReportView):
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
         context["personal"] = True
+        context["has_delete_permission"] = (self.request.user.account.
+                                            organization.has_delete_permission())
         return context
 
     def base_match_filter(self, reports):